Simple Roles Auth Control panel finished. Can not apply to conponents or routes to deny users without permissions. Also tidied up Newsfeed and user overview pages to give them a cleaner look and more responsive small screen look

+++ app/Http/Controllers/RoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Role;
 use Inertia\Inertia;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -15,9 +16,13 @@ class RoleController extends Controller
      */
     public function index()
     {
-        return Inertia::render('Admin/Roles/Index', [
-            'roles' => Role::all()
-        ]);
+        if (Gate::allows('accessRoles')) {
+
+            return Inertia::render('Admin/Roles/Index', [
+                'roles' => Role::all()
+            ]);
+        }
+        return back();
     }
 
     /**
@@ -27,7 +32,11 @@ class RoleController extends Controller
      */
     public function create()
     {
-        return Inertia::render('Admin/Roles/Create');
+        if (Gate::allows('manageRoles')) {
+
+            return Inertia::render('Admin/Roles/Create');
+        }
+        return back();
     }
 
     /**
@@ -38,15 +47,17 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required'
-        ]);
+        if (Gate::allows('manageRoles')) {
+            $request->validate([
+                'name' => 'required'
+            ]);
 
-        Role::create($request->only(['name']));
+            Role::create($request->only(['name']));
 
-        return redirect(route('AdminControlPanel.roles.index'));
-
-        //->withSuccess('Role created successfully!');
+            return redirect(route('AdminControlPanel.roles.index'));
+            //->withSuccess('Role created successfully!');
+        }
+        return back();
     }
 
     /**
@@ -57,9 +68,13 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
-        return Inertia::render('Admin/Roles/Edit', [
-            'role' => $role
-        ]);
+        if (Gate::allows('accessRoles')) {
+
+            return Inertia::render('Admin/Roles/Edit', [
+                'role' => $role
+            ]);
+        }
+        return back();
     }
 
     /**
@@ -82,16 +97,18 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
-        $request->validate([
-            'name' => 'required'
-        ]);
+        if (Gate::allows('manageRoles')) {
+            $request->validate([
+                'name' => 'required'
+            ]);
 
-        $role->update($request->only(['name']));
-
-        return redirect(route('AdminControlPanel.roles.index'));
-
-        //->withSuccess('Role updated successfully!');
+            $role->update($request->only(['name']));
+            return redirect(route('AdminControlPanel.roles.index'));
+            //->withSuccess('Role updated successfully!');
+        }
+        return back();
     }
+
 
     /**
      * Remove the specified resource from storage.
@@ -101,10 +118,11 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
-        $role->delete();
-
-        return redirect(route('AdminControlPanel.roles.index'));
-
-        //->withSuccess('Role deleted successfully!');
+        if (Gate::allows('manageRoles')) {
+            $role->delete();
+            return redirect(route('AdminControlPanel.roles.index'));
+            //->withSuccess('Role deleted successfully!');
+        }
+        return back();
     }
 }
